+++ R/sdplyr-Spatial.r
@@ -11,9 +11,12 @@ setOldClass( c("grouped_df", "tbl_df", "tbl", "data.frame" ) )
 #' @export
 #' @importFrom dplyr groups tbl_vars
 #' @examples 
+#' if (utils::packageVersion("dplyr") > "0.5.0") {
+#' 
 #' spmap %>% mutate_if(is.numeric, as.character)
 #' spmap %>% mutate_all(funs(as.character))
 #' spmap %>% mutate_at(vars(starts_with("L")), funs(as.integer))
+#' }
 tbl_vars.Spatial <- function(x) names(x)
 #' @name dplyr-S3
 #' @export